requirements.txt
Flask==1.1.4
Werkzeug==1.0.1
markupsafe==2.0.1

flask-restful
mysql-connector-python
flask-jwt-extended
haversine